Kenzie/SUB-MAF-14: Encapsulated notifications into an object class.
