+++ app/src/main/java/com.terracode.blueharvest/HomeActivity.kt
@@ -22,6 +22,7 @@ import com.terracode.blueharvest.utils.PreferenceManager
 import com.terracode.blueharvest.utils.ReadJSONObject
 import com.terracode.blueharvest.utils.UnitConverter
 import com.terracode.blueharvest.utils.viewManagers.LocaleManager
+import com.terracode.blueharvest.utils.viewManagers.NotificationManager
 import com.terracode.blueharvest.utils.viewManagers.TextSizeManager
 import com.terracode.blueharvest.utils.viewManagers.ThemeManager
 
@@ -132,7 +133,7 @@ class HomeActivity : AppCompatActivity() {
             R.id.notifications -> {
                 // Sample notifications (replace with your actual notifications)
                 val notifications = PreferenceManager.getNotifications()
-                showNotificationList(notificationButton, notifications)
+                NotificationManager.showNotificationList(this,notificationButton, notifications)
                 true
             }
             R.id.configurationSettings -> {
@@ -148,77 +149,4 @@ class HomeActivity : AppCompatActivity() {
             else -> false
         }
     }
-
-    private fun showNotificationList(anchorView: View, notifications: List<Notification>) {
-        val popupView = layoutInflater.inflate(R.layout.notification_layout, null)
-        val popupWindow = PopupWindow(popupView, WindowManager.LayoutParams.WRAP_CONTENT, WindowManager.LayoutParams.WRAP_CONTENT)
-
-        // Find the container layout in the popup view
-        val containerLayout = popupView.findViewById<LinearLayout>(R.id.notificationContainer)
-
-        // Find the clearNotificationsButton directly from the inflated view
-        val clearNotificationButton = popupView.findViewById<Button>(R.id.clearNotificationsButton)
-
-        // Add notifications dynamically to the container layout
-        if (notifications.isEmpty()){
-            val textView = TextView(this)
-            textView.text = "No Notifications Yet!"
-            containerLayout.addView(textView)
-        } else {
-            notifications.forEach { notification ->
-                val textView = TextView(this)
-                textView.text = "${notification.message}  ${notification.timestamp}"
-
-                // Set icon based on notification type
-                val icon = when (notification.type) {
-                    NotificationTypes.WARNING.toString() -> {
-                        val drawable =
-                            ContextCompat.getDrawable(this, R.drawable.exclamation_triangle_fill)
-                        drawable?.setTint(Color.YELLOW)
-                        drawable
-                    }
-
-                    NotificationTypes.ERROR.toString() -> {
-                        val drawable =
-                            ContextCompat.getDrawable(this, R.drawable.exclamation_circle_fill)
-                        drawable?.setTint(Color.RED)
-                        drawable
-                    }
-
-                    NotificationTypes.NOTIFICATION.toString() -> {
-                        val drawable = ContextCompat.getDrawable(this, R.drawable.info_circle_fill)
-                        drawable?.setTint(Color.BLUE)
-                        drawable
-                    }
-
-                    else -> null
-                }
-
-                // Set the drawable icon to the left of the text
-                icon?.let {
-                    textView.setCompoundDrawablesWithIntrinsicBounds(it, null, null, null)
-                    val padding = resources.getDimensionPixelSize(R.dimen.icon_padding)
-                    textView.compoundDrawablePadding = padding
-                }
-
-                containerLayout.addView(textView)
-            }
-        }
-
-        // Set up Listener for Clearing Notifications
-        clearNotificationButton.setOnClickListener {
-            // Call the method to clear notifications
-            PreferenceManager.clearNotifications()
-            // Dismiss the popup window after clearing notifications
-            popupWindow.dismiss()
-        }
-
-        // Set a dismiss listener to close the popup when clicked outside
-        popupWindow.isOutsideTouchable = true
-        popupWindow.isFocusable = true
-
-        // Show the popup below the anchor view (bell icon)
-        popupWindow.showAsDropDown(anchorView)
-    }
-
 }
